requirements.txt
Django==4.2.7
Pillow==10.1.0
django-extensions==3.2.3
python-decouple==3.8
gunicorn==21.2.0
whitenoise==6.6.0
psycopg2-binary==2.9.9
django-storages==1.14.2
boto3==1.34.0
django-cors-headers==4.3.1
djangorestframework==3.14.0
django-cleanup==8.0.0
exifread==3.0.0
iptcinfo3==2.1.4
piexif==1.1.3
pyexiftool==0.5.6
